(c-class-key, c-search-uplist-for-classkey): greatly simplified, this constant now contains a regexp that only matches valid `class' keywords, in the C++ sense. It now ignores any template, extern, static, etc. keywords.
This means there is a potential change in the relpos's for
class-open's and inclass's.  Instead of matching the boi of the line
containing the extern or template keyword, they return the boi of the
line containing the class, struct, or union keyword.

There're may be problems with this down the line, but for now, I think
this is much more winning than trying to hack that big regexp, which
never worked anyway.

--HG--
extra : convert_revision : svn%3Acb709983-ff18-4ef6-91a3-3ce3e3fe8e0b/trunk%40862