(c-forward-syntactic-ws): Added code that skips forward over multi-line cpp macros.
(c-beginning-of-macro): Moved, and made into a defsubst.  This
function can now actually find the beginning of a multi-line C
preprocessor macro.

(c-backward-syntactic-ws): Use c-beginning-of-macro to skip backwards
over multi-line macro definitions.

(c-in-literal, c-fast-in-literal): Use c-beginning-of-macro to find
out whether we're in a multi-line macro definition.

(c-guess-basic-syntax): CASE 3: we can now determine whether we're at
the beginning of a cpp macro definition, or inside the middle of one.
Set syntax to 'cpp-macro in the former case, 'cpp-macro-cont in the
latter.  In both cases, the relpos is the beginning of the macro.

--HG--
extra : convert_revision : svn%3Acb709983-ff18-4ef6-91a3-3ce3e3fe8e0b/trunk%401826