(c-offsets-alist): The offset for a syntactic symbol can now also be a list, which is recursively evaluated using the previous semantics. The first non-nil value succeeds.
(c-get-offset, c-evaluate-offset): Function to implement recursive
evaluation for when a symbol's offset is a list (the first one to
return a non-nil value succeeds -- should this be a stricter check for
an integer return value?).  Factored out of c-get-offset.

(c-set-offset, c-read-offset): Error message indicates value could be
a list.  In former function, add listp check.