(c-beginning-of-statement-1, c-crosses-statement-barrier-p): Rewritten to get a well defined and documented behavior.  This fixes some tricky cases in recognition of do-while constructs. #Got tired of the buggy behavior in c-beginning-of-statement-1 and all #the code that depends on it.  c-beginning-of-statement-1 now does a #better job and has a return value which probably can help to clean up #a lot of odd code in c-guess-*.
(c-backward-syntactic-ws): Fixed bug in skipping over a macro.

(c-backward-to-start-of-do, c-backward-to-start-of-if): Removed; use
c-beginning-of-statement-1 instead.

(c-guess-continued-construct, c-guess-basic-syntax): Various fixes to
not depend on the bugs previously in c-beginning-of-statement-1.
Cleanups in cases 18 and 17 to use the new behavior of
c-beginning-of-statement-1 better.  Fixed recognition of catch blocks
inside macros.

--HG--
extra : convert_revision : svn%3Acb709983-ff18-4ef6-91a3-3ce3e3fe8e0b/trunk%402883