(c-backward-labels-and-else): Removed again; c-add-stmt-syntax does this, but better.
(c-looking-at-bos): Added optional limit arg for backward searches.

(c-looking-at-inexpr-block): Anchor gcc in-expression statements at
the surrounding open parenthesis.  Treat a class body as an
in-expression class if it's used in an object clone expression in
Pike.

(c-guess-continued-construct, c-guess-basic-syntax): Improved and
unified anchoring at statements.  Statements beginning with comments
or labels are now avoided, by going out of blocks and containing
statements if necessary.  This nesting handling also fixes the case
when there's a statement after a block open brace.  Added the
syntactic symbol `substatement-label'.

(c-get-offset, c-get-syntactic-indentation): Allow several anchor
positions in the list of syntactic symbols.  Only the first is used as
the base for the offset calculation.

#(c-backward-to-block-anchor, c-add-stmt-syntax): Two new helper
#functions for the above.

#(c-beginning-of-statement-1): Bugfix in label handling.

#(c-add-class-syntax): Moved here for better locality.  Made it a
#function since we use dynamic binding for the syntax result in these
#areas anyway.