(c-syntactic-re-search-forward): Added an option to restrict matching to the top level of the current paren sexp.
(c-in-knr-argdecl): Broke out the K&R argdecl test in
`c-beginning-of-decl-1' to this new function.

(c-beginning-of-decl-1): Name change from
`c-beginning-of-declaration'.  Return more info.

(c-end-of-statement-1, c-end-of-decl-1): Replaced
`c-end-of-statement-1' with `c-end-of-decl-1', which correctly handles
declarations that continue after the block.

--HG--
extra : convert_revision : svn%3Acb709983-ff18-4ef6-91a3-3ce3e3fe8e0b/trunk%403085