Add the `auto-mode-alist' entries to the end instead of the start of the list, to avoid overriding user additions that are made before this file is loaded when the entries don't correspond to any already on `auto-mode-alist'.
This is done through the third append argument to `auto-mode-alist'.
That doesn't exist in older (X)Emacsen, so the function is adviced in
cc-defs.  However, that advice doesn't help if the autoloads are
updated in an old (X)Emacs with this version of CC Mode, but I believe
it's unlikely that anyone does that when CC Mode isn't distributed
with with it.

--HG--
extra : convert_revision : svn%3Acb709983-ff18-4ef6-91a3-3ce3e3fe8e0b/trunk%403623