cc-langs.el (c-type-decl-prefix-key): C++ bit: move "\(const\|throw\|volatile\)\>" nearer the start of the regexp, so that these keywords aren't wrongly matched as identifiers.
cc-fonts.el (c-font-lock-declarators): In (match-beginning 2), "2" ->
"3", corresponding to the change in c-type-decl-prefix-key.
(c-font-lock-maybe-decl-faces): Rearranged the header comments.

cc-engine.el (c-find-decl-spots): Rearranged the header comments.
(c-forward-decl-or-cast-1): (i) In (match-beginning 2), "2" -> "3",
corresponding to the change in c-type-decl-prefix-key.  (ii) Remove a
check for a semicolon at the end of a function declaration: this causes
C++ constructor declarations to be fontified before the semicolon is
typed.